fix($identify): do not alias same anon id to multiple identified users
This arises when we have this sequence of events:

```
posthog.capture('event1')
posthog.identify('user1')
posthog.capture('event2')
posthog.identify('user2')
posthog.capture('event3')
```

We want to ensure we associate the anonymous 'event1' with 'user1' when
they are identified, but we do not want to alias 'user1' and 'user2' as
these are separate users. With the previous logic we would associate
both 'user1' and 'user2' with the anonymous id associated with 'event1'
and therefore by transitivity, 'user1' and 'user2' would be associated.

Closes https://github.com/PostHog/posthog/issues/5527